Added initial project files
projectHome.sh as it stands is a shell script that takes input from config.env to get the Google Cloud storage location of the files containing the data from the sequences to be analyzed. It parses the file list to obtain a list of chromosome identifiers used in the folder, then is capable of downloading these files.